+++ src/main/java/docsite/emitters/MarkdownGeneratedSectionEmitter.java
@@ -4,14 +4,23 @@ package docsite.emitters;
 import java.io.*;
 import java.util.stream.*;
 
-import com.vladsch.flexmark.ast.Heading;
-import com.vladsch.flexmark.util.ast.Node;
+import com.vladsch.flexmark.ast.*;
+import com.vladsch.flexmark.html.HtmlRenderer;
+import com.vladsch.flexmark.parser.Parser;
+import com.vladsch.flexmark.util.ast.*;
+import com.vladsch.flexmark.util.data.MutableDataSet;
 import docsite.*;
 import docsite.util.ResourceUtil;
 import j2html.tags.specialized.*;
 import static j2html.TagCreator.*;
 
 public class MarkdownGeneratedSectionEmitter extends GeneratedSectionEmitter {
+
+
+    protected static final MutableDataSet options = new MutableDataSet();
+    protected static final Parser parser = Parser.builder(options).build();
+    protected static final HtmlRenderer renderer = HtmlRenderer.builder(options).build();
+
 
     public MarkdownGeneratedSectionEmitter(EmitterBuildParams params) {
         super(params);