+++ CHANGELOG.md
@@ -7,12 +7,23 @@ The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
 
+[1.1.0]
+--------------------------------------------------------------------------
+**Release date:** 2022-01-06
+
+### Added
+- Added `metadata` and `scripts` elements to the site configuration
+(useful for adding analytic scripts)
+- Added `favicon` to the site configuration
+- Added a *jump to content* button when navigating with the `Tab` key
+- HTML `blockquote` tags now have differentiated style
+
 [1.0.0] 
 --------------------------------------------------------------------------
-**Release date:** 2021-08-11
+**Release date:** 2021-12-17
 
 - Initial release.  
 
 
-[1]: <https://keepachangelog.com/en/1.0.0/>
+[1]: <https://keepachangelog.com>
 [2]: <https://semver.org/>